pyproject.toml
[tool.poetry]
name = "news-assistant-ai"
version = "0.1.0"
description = ""
authors = ["Mikhail Makarov <vallome@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.13"
langchain-community = "^0.3.13"
langchain-openai = "^0.2.14"
scikit-learn = "^1.6.0"
faiss-cpu = "^1.9.0.post1"
langgraph = "^0.2.60"
streamlit = "^1.41.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
